backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
PyMySQL==1.1.1
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-jose==3.3.0
pydantic==2.9.2
email-validator==2.2.0
python-multipart==0.0.9
numpy==2.1.3
scikit-learn==1.5.2
tensorflow==2.20.0
opencv-python==4.10.0.84
pdfminer.six==20231228
pandas==2.2.3
